refactor(acp): respond to PR review
A handful of focused cleanups, no behaviour change.

- daemon.py: drop the "Inputs are container-shaped" / "subprocess vs
  in-process" prose from the module docstring — defining what we're
  *not* rather than what we are, awkward for an API-shaped module
  whose signature already speaks for itself.

- probe.py: drop the deprecated ``loop=`` kwargs on
  ``StreamReader`` / ``StreamReaderProtocol``.  Modern asyncio
  resolves the running loop automatically.

- roster.py: quote the agent name in the wrapper-existence shell
  script — defensive against agent names that pick up weird chars
  (e.g. globs) in some future config.

- model_options.py: pull the four model-frame helpers out of
  proxy.py and into the existing model-options vocabulary module
  (``_humanise_model_id``, ``_build_model_config_option``,
  ``_build_session_new_response``, ``_rewrite_model_options_in_place``).
  proxy.py drops from 921 → 792 lines; model_options.py grows to
  match.  Generic helpers (``_summarise_frame``,
  ``_with_params_field``) stay in proxy.py since they're not
  model-specific.

- claude.yaml: tighten the symlink-step comment so its rationale
  doesn't read as conflated with the cwd-rewrite stopgap (the
  symlink fixes an SDK-load check; cwd is a separate spawn-time
  issue).

- tests: drop the dead ``_read_request_id`` helper from
  test_acp_probe.py; update the model-helper imports in
  test_acp_proxy.py to point at model_options.py; update the
  shell-script fixture in test_acp_roster.py to match the new
  quoting.
